renamed modell to model; moved com.evrencoskun.tableviewsample.holder to com.evrencoskun.tableviewutil.holder
