+++ app/src/main/java/com/evrencoskun/tableviewsample/tableview/TableViewAdapterBase.java
@@ -38,10 +38,10 @@ import com.evrencoskun.tableview.sort.ISortableModel;
 import com.evrencoskun.tableview.sort.SortState;
 import com.evrencoskun.tableviewsample.R;
 import com.evrencoskun.tableviewsample.tableview.holder.ColumnHeaderViewHolder;
-import com.evrencoskun.tableviewsample.tableview.holder.GenericTextCellViewHolder;
-import com.evrencoskun.tableviewsample.tableview.holder.RowHeaderViewHolder;
-import com.evrencoskun.tableview.modell.Cell;
-import com.evrencoskun.tableviewsample.tableview.model.RowHeader;
+import com.evrencoskun.tableviewutil.holder.GenericTextCellViewHolder;
+import com.evrencoskun.tableviewutil.holder.RowHeaderViewHolder;
+import com.evrencoskun.tableview.model.Cell;
+import com.evrencoskun.tableview.model.RowHeader;
 
 /**
  * Generic adapter that translates a Pojo to TableView-Cells.
