+++ tableview/src/main/java/com/evrencoskun/tableview/model/IColumnValue.java
@@ -1,4 +1,4 @@
-package com.evrencoskun.tableview.modell;
+package com.evrencoskun.tableview.model;
 
 public interface IColumnValue<T> {
     Object get(T row);
