+++ tableviewutil/build.gradle
@@ -30,6 +30,7 @@ dependencies {
     implementation project(path: ':tableview')
 
     implementation "androidx.appcompat:appcompat:$androidx_appcompat_version"
+    implementation "androidx.recyclerview:recyclerview:$androidx_recyclerview_version"
     testImplementation "junit:junit:$junit_version"
     androidTestImplementation "androidx.test.ext:junit:$androidx_test_ext_version"
     androidTestImplementation "androidx.test:rules:$androidx_test_version"
